requirements.txt
gdown==4.7.1
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.11.2
oauth2client==4.1.3
pandas==2.1.1
python-dotenv==1.0.0